+++ VertexCover/VertexCover/Src/KernelizedAttributes.cs
